+++ src/errorHandlers/handleHttpClientErrorResponse.ts
@@ -14,6 +14,14 @@ export async function handleHttpClientErrorResponse(error: HttpErrorResponse, op
     //handles client codes (400..499)
     if(error.status >= 400 && error.status < 500)
     {
+        const ignoredCodes = options.ignoredHttpStatusCodes ?? [];
+
+        //ignored code, not handled
+        if(ignoredCodes.indexOf(error.status) >= 0)
+        {
+            return error;
+        }
+
         const {errors, validationErrors} = await readErrorsFromHttpErrorResponse(error, options.injector, options.clientErrorsResponseMapper, options.clientValidationErrorsResponseMapper);
 
         return new HttpClientError(errors, validationErrors, error);